Fix hotkey router entry bindings
